+++ notification-pusher/notification-pusher-actionmailer/app/mailers/notification_pusher/action_mailer/notification_mailer.rb
@@ -3,8 +3,10 @@
 module NotificationPusher
   class ActionMailer
     class NotificationMailer < ApplicationMailer
+      # rubocop:disable Metrics/ParameterLists
       def push(notification,
-               from:, to: nil, renderer: 'actionmailer', layout: nil, mail_options: {})
+               from:, to: nil, renderer: 'actionmailer', layout: nil,
+               mail_options: {})
         render(layout: layout) unless layout.nil?
 
         @notification = notification
@@ -15,6 +17,7 @@ module NotificationPusher
           from: from
         }.merge(mail_options))
       end
+      # rubocop:enable Metrics/ParameterLists
     end
   end
 end