+++ notification-pusher/notification-pusher-actionmailer/lib/notification_pusher/delivery_method/action_mailer.rb
@@ -5,21 +5,25 @@ module NotificationPusher
     class ActionMailer < NotificationPusher::DeliveryMethod::Base
       require_relative 'action_mailer/engine'
 
-      DEFAULT_MAILER_CLASS = NotificationPusher::ActionMailer::NotificationMailer.freeze
-      DEFAULT_MAILER_ACTION = :push.freeze
-      DEFAULT_DELIVER_METHOD = :deliver.freeze
+      DEFAULT_MAILER_ACTION = :push
+      DEFAULT_DELIVER_METHOD = :deliver
 
       def call
-        mailer_class.send(mailer_action, notification, options).send(deliver_method)
+        mailer_class.send(mailer_action, notification, options)
+                    .send(deliver_method)
       end
 
       private
 
       def mailer_class
         return options[:mailer] if options.key?(:mailer)
-        return DEFAULT_MAILER_CLASS if defined?(NotificationRenderer)
+        if defined?(NotificationRenderer)
+          return ::NotificationPusher::ActionMailer::NotificationMailer
+        end
 
-        raise(ArgumentError, "You have to pass the :mailer option explicitly or require 'notification-renderer'")
+        raise(ArgumentError,
+              'You have to pass the :mailer option explicitly or require ' \
+              "'notification-renderer'")
       end
 
       def mailer_action